+++ CHANGELOG.md
@@ -26,3 +26,41 @@ All notable changes to this project will be documented in this file.
 - Supports inline and column layouts
 - Customizable colors, borders, icons
 - Demo app included in example folder
+
+
+## [1.0.3] 
+- Initial stable release
+- Supports inline and column layouts
+- Customizable colors, borders, icons
+- Demo app included in example folder
+
+
+## [1.0.4] 
+- Initial stable release
+- Supports inline and column layouts
+- Customizable colors, borders, icons
+- Demo app included in example folder
+
+
+## [1.0.5] 
+- Initial stable release
+- Supports inline and column layouts
+- Customizable colors, borders, icons
+- Demo app included in example folder
+
+
+
+## [1.0.6] 
+- Initial stable release
+- Supports inline and column layouts
+- Customizable colors, borders, icons
+- Demo app included in example folder
+
+
+## [1.0.7]
+- Initial stable release
+- Supports inline, column, and new suggestion layouts: `insideBelow`, `outsideAbove`, `outsideBelow`
+- Added support for showing suggestions/tags in flexible positions
+- Customizable colors, borders, icons
+- Fixed tag overflow issues
+- Demo app included in the example folder